+++ src/react/hooks/useLazyQuery.ts
@@ -2,13 +2,15 @@ import { DocumentNode } from 'graphql';
 import { TypedDocumentNode } from '@graphql-typed-document-node/core';
 import { useCallback, useMemo, useState } from 'react';
 
+import { OperationVariables } from '../../core';
 import {
   LazyQueryHookOptions,
   QueryLazyOptions,
   QueryTuple,
+  LazyQueryHookOptionsFunction,
 } from '../types/types';
 import { useQuery } from './useQuery';
-import { OperationVariables } from '../../core';
+import { useNormalizedOptions } from './options';
 
 // The following methods, when called will execute the query, regardless of
 // whether the useLazyQuery execute function was called before.
@@ -22,8 +24,12 @@ const EAGER_METHODS = [
 
 export function useLazyQuery<TData = any, TVariables = OperationVariables>(
   query: DocumentNode | TypedDocumentNode<TData, TVariables>,
-  options?: LazyQueryHookOptions<TData, TVariables>
+  optionsOrFunction?:
+    | LazyQueryHookOptions<TData, TVariables>
+    | LazyQueryHookOptionsFunction<TData, TVariables>
 ): QueryTuple<TData, TVariables> {
+  const options = useNormalizedOptions(optionsOrFunction);
+
   const [execution, setExecution] = useState<{
     called: boolean,
     options?: QueryLazyOptions<TVariables>,